I have moved the connnection manager tests to test package and added temperature utility to utilities package
