--- src/test/java/com/sparta/sc/App.java
@@ -1,4 +1,0 @@
-package com.sparta.sc;
-
-public class App {
-}
